+++ src/utils/buffer.ts
@@ -47,3 +47,34 @@ export const bufferToString = (buffer: ArrayBuffer): string => {
   }
   return buffer
 }
+
+const _decodeURIComponent = decodeURIComponent
+
+export const bufferToFormData = (arrayBuffer: ArrayBuffer, contentType: string) => {
+  const decoder = new TextDecoder('utf-8')
+  const content = decoder.decode(arrayBuffer)
+  const formData = new FormData()
+
+  const boundaryMatch = contentType.match(/boundary=(.+)/)
+  const boundary = boundaryMatch ? boundaryMatch[1] : ''
+
+  if (contentType.startsWith('multipart/form-data') && boundary) {
+    const parts = content.split('--' + boundary).slice(1, -1)
+    for (const part of parts) {
+      const [header, body] = part.split('\r\n\r\n')
+      const nameMatch = header.match(/name="([^"]+)"/)
+      if (nameMatch) {
+        const name = nameMatch[1]
+        formData.append(name, body.trim())
+      }
+    }
+  } else if (contentType.startsWith('application/x-www-form-urlencoded')) {
+    const pairs = content.split('&')
+    for (const pair of pairs) {
+      const [key, value] = pair.split('=')
+      formData.append(_decodeURIComponent(key), _decodeURIComponent(value))
+    }
+  }
+
+  return formData
+}
